backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
openai==1.12.0
python-docx==1.1.0
pypdf2==3.0.1
pdfplumber==0.11.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
weasyprint==61.2
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.2.1
